修改代码

+++ src/main/java/com/spider/controller/FileController.java
@@ -1,9 +1,11 @@
 package com.spider.controller;
 
+import com.spider.utils.ExchangeUtil;
 import com.spider.utils.FileUtils;
 import com.spider.vo.ResponseVo;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -12,7 +14,6 @@ import org.springframework.web.bind.annotation.RestController;
 import java.io.File;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 
 @Api(tags = "文件接口")
@@ -42,40 +43,47 @@ public class FileController extends BaseController {
             System.out.println("-----------------------------------------");
             Map<String, Object> map = new HashMap<>();
             AtomicInteger keyCount = new AtomicInteger(0);
-            List<String> textList = new ArrayList<>();
-            textList.addAll(FileUtils.readTxt(book.getAbsolutePath(), "GB18030"));
-            textList.addAll(FileUtils.readTxt(book.getAbsolutePath(), "GBK"));
-            textList.addAll(FileUtils.readTxt(book.getAbsolutePath(), "GB2312"));
-            textList.addAll(FileUtils.readTxt(book.getAbsolutePath(), "ASCII"));
-            textList.addAll(FileUtils.readTxt(book.getAbsolutePath(), "UTF-8"));
-            textList.addAll(FileUtils.readTxt(book.getAbsolutePath(), "BIG5"));
-            List<String> matchKey = new ArrayList<>();
-            List<String> matchLine = new ArrayList<>();
-            textList.stream().distinct().forEach(line -> {
-                keyList.stream().distinct().filter(key -> !key.startsWith("#")).forEach(key -> {
-                    if (line.contains(key)) {
-                        keyCount.incrementAndGet();
-                        matchKey.add(key);
-                        matchLine.add(line);
-                        System.out.println(book.getName() + "," + key + "," + line);
-                    }
+            String charset = getTextCharset(book);
+            if (!StringUtils.isBlank(charset)) {
+                List<String> textList = FileUtils.readTxt(book.getAbsolutePath(), charset);
+                List<String> matchKey = new ArrayList<>();
+                List<String> matchLine = new ArrayList<>();
+                textList.stream().distinct().forEach(line -> {
+                    keyList.stream().distinct().filter(key -> !key.startsWith("#")).forEach(key -> {
+                        if (line.contains(key)) {
+                            keyCount.incrementAndGet();
+                            matchKey.add(key);
+                            matchLine.add(line);
+                            System.out.println(book.getName() + "," + key + "," + line);
+                        }
+                    });
                 });
-            });
-            if (keyCount.get() > matchCount) {
-                map.put("savePath", book.getAbsolutePath());
-                map.put("keyCount", keyCount.get());
-                map.put("matchKey", new HashSet<>(matchKey));
-                map.put("matchLine", new HashSet<>(matchLine));
-                if (new File(targetPath + book.getName()).exists()) {
-                    book.delete();
-                } else {
-                    book.renameTo(new File(targetPath + book.getName()));
+                if (keyCount.get() > matchCount) {
+                    map.put("savePath", book.getAbsolutePath());
+                    map.put("keyCount", keyCount.get());
+                    map.put("matchKey", new HashSet<>(matchKey));
+                    map.put("matchLine", new HashSet<>(matchLine));
+                    if (new File(targetPath + book.getName()).exists()) {
+                        book.delete();
+                    } else {
+                        book.renameTo(new File(targetPath + book.getName()));
+                    }
+                    return map;
                 }
-                return map;
+                return null;
             }
             return null;
         }).filter(Objects::nonNull).collect(Collectors.toList());
         return ResponseVo.succee(result);
     }
 
+    public String getTextCharset(File file) {
+        try {
+            ExchangeUtil exchangeUtil = new ExchangeUtil();
+            return ExchangeUtil.javaname[exchangeUtil.detectEncoding(file)];
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
+        }
+    }
 }
